+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## Changelog
 
 ### 0.0.7-dev-pre
+
 - Pagination api
-- 
+- Pagination frontend
 
 ### 0.0.6-dev
+
 - Autosave articles
 - Now non-writers cannot access the write page
 - Profile preview